// frontend/src/app/empleado/empleado.types.ts
export interface Empleado {
  id: number;
  nombreCompleto: string;
  telefono: string;
  dpi: string;
  cv: any;
  fotografia: string;
  domicilio: string;
  edad: number;
  genero: string;
  seguroSocial: string;
  cuentaBancaria: string;
  estadoCivilId: string;
  email: string;
  codigoEmpleado: string;
  password: string;
  estado: string;
}